+++ DataObjects/Tags.cs
@@ -4,33 +4,33 @@
 	/// Represents a tag entity used for categorizing stories, including its text and ban status.
 	/// </summary>
 	public record Tag(
-		[property: JsonPropertyName("id")] int? Id,
+		[property: JsonPropertyName("id")] long? Id,
 		[property: JsonPropertyName("tag"), JsonConverter(typeof(StringOrIntConverter))] StringOrInt TagText,
-		[property: JsonPropertyName("is_banned")] int? IsBanned
+		[property: JsonPropertyName("is_banned")] long? IsBanned
 	);
 
 	/// <summary>
 	/// Represents detailed information about a category, including its name, descriptions, and submission count.
 	/// </summary>
 	public record CategoryReturn(
-		[property: JsonPropertyName("id")] int? Id,
-		[property: JsonPropertyName("language")] int? Language,
+		[property: JsonPropertyName("id")] long? Id,
+		[property: JsonPropertyName("language")] long? Language,
 		[property: JsonPropertyName("ldesc")] string Ldesc,
 		[property: JsonPropertyName("name")] string Name,
 		[property: JsonPropertyName("pageUrl")] string PageUrl,
 		[property: JsonPropertyName("sdesc")] string Sdesc,
 		[property: JsonPropertyName("topUrl")] string TopUrl,
 		[property: JsonPropertyName("type")] string Type,
-		[property: JsonPropertyName("submission_count")] int? SubmissionCount
+		[property: JsonPropertyName("submission_count")] long? SubmissionCount
 	);
 
 	/// <summary>
 	/// Represents a tag that is related to another tag, including its ID, display text, and usage count.
 	/// </summary>
 	public record RelatedTag(
-		[property: JsonPropertyName("id")] int? Id,
+		[property: JsonPropertyName("id")] long? Id,
 		[property: JsonPropertyName("tag"), JsonConverter(typeof(StringOrIntConverter))] StringOrInt Tag,
-		[property: JsonPropertyName("cnt")] int? Cnt
+		[property: JsonPropertyName("cnt")] long? Cnt
 	);
 
 	/// <summary>
@@ -44,10 +44,10 @@
 	/// Represents available statistical checks across multiple time periods such as week, month, and today.
 	/// </summary>
 	public record PeriodChecks(
-		[property: JsonPropertyName("month")] int? Month,
-		[property: JsonPropertyName("week")] int? Week,
-		[property: JsonPropertyName("today")] int? Today,
-		[property: JsonPropertyName("allperiod")] int? Allperiod
+		[property: JsonPropertyName("month")] long? Month,
+		[property: JsonPropertyName("week")] long? Week,
+		[property: JsonPropertyName("today")] long? Today,
+		[property: JsonPropertyName("allperiod")] long? Allperiod
 	);
 
 	/// <summary>
@@ -62,12 +62,12 @@
 	/// Represents detailed information about a specific tag, including category, language, and usage metrics.
 	/// </summary>
 	public record TagInfo(
-		[property: JsonPropertyName("id")] int? Id,
+		[property: JsonPropertyName("id")] long? Id,
 		[property: JsonPropertyName("tag")] string Tag,
-		[property: JsonPropertyName("language")] int? Language,
-		[property: JsonPropertyName("tagid")] int? Tagid,
-		[property: JsonPropertyName("category")] int? Category,
-		[property: JsonPropertyName("cnt")] int? Cnt,
-		[property: JsonPropertyName("is_banned")] int? IsBanned
+		[property: JsonPropertyName("language")] long? Language,
+		[property: JsonPropertyName("tagid")] long? Tagid,
+		[property: JsonPropertyName("category")] long? Category,
+		[property: JsonPropertyName("cnt")] long? Cnt,
+		[property: JsonPropertyName("is_banned")] long? IsBanned
 	);
 }
